+++ src/cli/main.rs
@@ -1,10 +1,12 @@
 use anyhow::{format_err, Result};
+use ignore::{overrides::OverrideBuilder, WalkBuilder};
 use std::fs;
-use std::path::{Path, PathBuf};
 use std::io::{stdin, stdout, Read, Write};
-use structopt::StructOpt;
+use std::path::{Path, PathBuf};
+use structopt::{clap::arg_enum, StructOpt};
 use stylua_lib::{format_code, Config};
-use ignore::{WalkBuilder, overrides::OverrideBuilder};
+
+mod output_diff;
 
 #[derive(StructOpt, Debug)]
 #[structopt(name = "stylua", about = "A utility to format Lua code")]
@@ -12,6 +14,22 @@ struct Opt {
     /// Specify path to stylua.toml
     #[structopt(long = "config-path", parse(from_os_str))]
     config_path: Option<PathBuf>,
+
+    /// Runs in 'check' mode
+    /// Exits with 0 if all formatting is OK
+    /// Exits with 1 if the formatting is incorrect
+    /// Any files input will not be overwritten
+    #[structopt(short, long)]
+    check: bool,
+
+    // Whether the output should include terminal colour or not
+    #[structopt(
+        long,
+        possible_values = &Color::variants(),
+        case_insensitive = true,
+        default_value = "auto",
+    )]
+    color: Color,
 
     /// A glob pattern to test against which files to check
     #[structopt(long, default_value = "**/*.lua")]
@@ -22,7 +40,26 @@ struct Opt {
     files: Vec<PathBuf>,
 }
 
-fn format_file(path: &Path, config: Config) -> Result<()> {
+structopt::clap::arg_enum! {
+    #[derive(Clone, Copy, Debug, PartialEq)]
+    pub enum Color {
+        Always,
+        Auto,
+        Never,
+    }
+}
+
+impl Color {
+    /// Whether we should use a coloured terminal.
+    pub fn use_colored_tty(self) -> bool {
+        match self {
+            Color::Always | Color::Auto => true,
+            Color::Never => false,
+        }
+    }
+}
+
+fn format_file(path: &Path, config: Config, check_only: bool, color: Color) -> Result<i32> {
     match fs::read(path) {
         Ok(contents) => {
             let contents = String::from_utf8_lossy(&contents);
@@ -37,13 +74,27 @@ fn format_file(path: &Path, config: Config) -> Result<()> {
                 }
             };
 
-            match fs::write(path, formatted_contents) {
-                Ok(_) => Ok(()),
-                Err(error) => Err(format_err!(
-                    "error: could not write to file {}: {}",
-                    path.display(),
-                    error
-                )),
+            if check_only {
+                let diff = output_diff::make_diff(&contents, &formatted_contents, 3);
+                if diff.is_empty() {
+                    Ok(0)
+                } else {
+                    output_diff::print_diff(
+                        diff,
+                        |line| format!("Diff in {} at line {}:", path.display(), line),
+                        color,
+                    );
+                    Ok(1)
+                }
+            } else {
+                match fs::write(path, formatted_contents) {
+                    Ok(_) => Ok(0),
+                    Err(error) => Err(format_err!(
+                        "error: could not write to file {}: {}",
+                        path.display(),
+                        error
+                    )),
+                }
             }
         }
         Err(error) => Err(format_err!(
@@ -60,25 +111,14 @@ fn format_string(input: String, config: Config) -> Result<()> {
     let out = &mut stdout();
     let formatted_contents = match format_code(&input, config) {
         Ok(formatted) => formatted,
-        Err(error) => {
-            return Err(format_err!(
-                "error: could not format from stdin: {}",
-                error
-            ))
-        }
+        Err(error) => return Err(format_err!("error: could not format from stdin: {}", error)),
     };
 
     match out.write_all(&formatted_contents.into_bytes()) {
         Ok(()) => Ok(()),
-        Err(error) => {
-            return Err(format_err!(
-                "error: could not output to stdout: {}",
-                error
-            ))
-        }
+        Err(error) => return Err(format_err!("error: could not output to stdout: {}", error)),
     }
 }
-
 
 fn format(opt: Opt) -> Result<i32> {
     if opt.files.is_empty() {
@@ -117,6 +157,7 @@ fn format(opt: Opt) -> Result<i32> {
     };
 
     let mut errors = vec![];
+    let mut error_code = 0;
 
     for file_path in opt.files.iter() {
         let override_builder = match OverrideBuilder::new(file_path).add(&opt.pattern) {
@@ -127,36 +168,53 @@ fn format(opt: Opt) -> Result<i32> {
                     opt.pattern,
                     error
                 ));
-                continue
-            },
-        }.build().unwrap(); // TODO: don't unwrap like this
+                continue;
+            }
+        }
+        .build()
+        .unwrap(); // TODO: don't unwrap like this
 
-        let walker = WalkBuilder::new(file_path).standard_filters(false).hidden(true).overrides(override_builder).build();
+        let walker = WalkBuilder::new(file_path)
+            .standard_filters(false)
+            .hidden(true)
+            .overrides(override_builder)
+            .build();
 
         for result in walker {
             match result {
                 Ok(entry) => {
                     if entry.is_stdin() {
+                        if opt.check {
+                            errors.push(format_err!(
+                                "warning: `--check` cannot be used whilst reading from stdin"
+                            ))
+                        };
+
                         let mut buf = String::new();
                         match stdin().read_to_string(&mut buf) {
                             Ok(_) => match format_string(buf, config) {
                                 Ok(_) => continue,
-                                Err(error) => errors.push(error)
+                                Err(error) => errors.push(error),
                             },
-                            Err(error) => errors.push(format_err!("error: could not read from stdin: {}", error))
+                            Err(error) => errors
+                                .push(format_err!("error: could not read from stdin: {}", error)),
                         }
                     } else {
                         let path = entry.path();
                         if path.is_file() {
-                            match format_file(path, config) {
-                                Ok(_) => continue,
-                                Err(error) => errors.push(error)
+                            match format_file(path, config, opt.check, opt.color) {
+                                Ok(code) => {
+                                    if code != 0 {
+                                        error_code = code
+                                    }
+                                }
+                                Err(error) => errors.push(error),
                             }
                         } else {
                             errors.push(format_err!("error: {} is not a file", path.display()));
                         }
                     }
-                },
+                }
                 Err(error) => {
                     errors.push(format_err!("error: could not walk: {}", error));
                 }
@@ -171,7 +229,7 @@ fn format(opt: Opt) -> Result<i32> {
         return Ok(1);
     }
 
-    Ok(0)
+    Ok(error_code)
 }
 
 fn main() {